refactor: hooksとcontextディレクトリの場所を変更_#25
