+++ pages/_app.js
@@ -1,8 +1,8 @@
 import "../components/global.css";
-import PageStates from "../components/context/PageStates";
-import PostStates from "../components/context/PostStates";
-import CommentStates from "../components/context/CommentStates";
-import LikeStates from "../components/context/LikeStates";
+import PageStates from "../context/PageStates";
+import PostStates from "../context/PostStates";
+import CommentStates from "../context/CommentStates";
+import LikeStates from "../context/LikeStates";
 
 export default function App({ Component, pageProps }) {
   return (
